working tests and refactor import online
